renamend package shared -> base
